feat: add a serve function
