+++ datafusion-postgres/Cargo.toml
@@ -25,4 +25,4 @@ log = "0.4"
 pgwire = { workspace = true }
 postgres-types = "0.2"
 rust_decimal = { version = "1.37", features = ["db-postgres"] }
-tokio = { version = "1.45", features = ["sync"] }
+tokio = { version = "1.45", features = ["sync", "net"] }
